refactored controllers and added generic view-profile to prep for room modifications
