+++ src/main/java/com/reservation/hotel/HotelReservation/config/LoginSuccessHandler.java
@@ -33,7 +33,7 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
                     log.info("Guest log in success!");
                     String query = user.getUsername();
 //                    String url = "/guest-profile" + "?username=" + query;
-                    String url = "/guest-profile";
+                    String url = "/profile";
                     redirectStrategy.sendRedirect(request, response, url);
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -44,7 +44,8 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
                     log.info("{}", user);
                     log.info("Clerk log in success!");
                     String query = user.getUsername();
-                    String url = "/clerk-profile" + "?username=" + query;
+//                    String url = "/clerk-profile" + "?username=" + query;
+                    String url = "/profile";
                     redirectStrategy.sendRedirect(request, response, url);
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -54,9 +55,10 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
                     // TODO: update to point to an admin profile page instead of clerk profile
                     User user = (User) authentication.getPrincipal();
                     log.info("{}", user);
-                    log.info("Clerk log in success!");
+                    log.info("Admin log in success!");
                     String query = user.getUsername();
-                    String url = "/clerk-profile" + "?username=" + query;
+//                    String url = "/clerk-profile" + "?username=" + query;
+                    String url = "/profile";
                     redirectStrategy.sendRedirect(request, response, url);
                 } catch (Exception e) {
                     e.printStackTrace();
