+++ src/main/resources/templates/view-profile.html
@@ -1,0 +1,147 @@
+<!doctype html>
+<html lang="en" xmlns:th="https://www.thymeleaf.org">
+<head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+
+    <title>Hello, world!</title>
+</head>
+<body>
+
+<nav class="navbar navbar-expand-sm navbar-light" style="background-color: #e3f2fd;">
+	<a class="navbar-brand" href="#">Employee Profile Page</a>
+	<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+		<span class="navbar-toggler-icon"></span>
+	</button>
+
+	<div class="collapse navbar-collapse" id="navbarSupportedContent">
+		<ul class="navbar-nav mr-auto">
+			<li class="nav-item active">
+				<a class="nav-link" th:href="@{/}">Home</a>
+			</li>
+			<!--            <li class="nav-item">-->
+			<!--                <a class="nav-link" href="#">Link</a>-->
+			<!--            </li>-->
+		</ul>
+		<span class="form-inline my-2 my-lg-0">
+            <div class="btn-group mr-2" role="group" aria-label="Logout">
+                <a th:href="@{/logout}" class="btn btn-light active" role="button" aria-pressed="true">Logout</a>
+            </div>
+        </span>
+	</div>
+</nav>
+
+<section class="m-4" th:object="${hotelUser}">
+
+        <div class="row" >
+            <div class="col-lg-4">
+                <div class="card mb-4" style="background-color: #e3f2fd;">
+                    <div class="card-body text-center">
+                        <img src="https://mdbcdn.b-cdn.net/img/Photos/new-templates/bootstrap-chat/ava3.webp" alt="avatar"
+                             class="rounded-circle img-fluid" style="width: 150px;">
+                        <h5 class="my-3" th:text="*{firstName +' '+ lastName}"></h5>
+                        <div class="d-flex justify-content-center mb-2">
+	                        <form th:action="@{/edit-profile}" method="get">
+		                        <button type="submit" class="btn btn-info">Edit Profile</button>
+	                        </form>
+                        </div>
+	                    <div th:if="${hotelUser.role == 'ROLE_GUEST'}" class="d-flex justify-content-center mb-2">
+		                    <button type="button" class="btn btn-info">View my reservations</button>
+	                    </div>
+	                    <div class="d-flex justify-content-center mb-2">
+	                        <form th:action="@{/rooms}" method="get">
+		                        <button type="submit" class="btn btn-info">View All Rooms</button>
+	                        </form>
+                        </div>
+	                    <div th:if="${hotelUser.role == 'ROLE_ADMIN'}" class="d-flex justify-content-center mb-2">
+		                    <form th:action="@{/admin}" method="get">
+			                    <button type="submit" class="btn btn-info">Admin Dashboard</button>
+		                    </form>
+	                    </div>
+                    </div>
+                </div>
+            </div>
+            <div class="col-lg-8">
+                <div class="card mb-4" style="background-color: #e3f2fd;">
+                    <div class="card-body">
+                        <div class="row">
+                            <div class="col-sm-3">
+                                <p class="mb-0">Full Name</p>
+                            </div>
+                            <div class="col-sm-9">
+                                <p class="text-muted mb-0" th:text="*{firstName +' '+ lastName}"></p>
+                            </div>
+                        </div>
+                        <hr>
+                        <div class="row">
+                            <div class="col-sm-3">
+                                <p class="mb-0">Email</p>
+                            </div>
+                            <div class="col-sm-9">
+                                <p class="text-muted mb-0" th:text="*{email}"></p>
+                            </div>
+                        </div>
+                        <hr>
+                        <div class="row">
+                            <div class="col-sm-3">
+                                <p class="mb-0">Phone</p>
+                            </div>
+                            <div class="col-sm-9">
+                                <p class="text-muted mb-0" th:text="*{phoneNumber}"></p>
+                            </div>
+                        </div>
+                        <hr>
+                        <div class="row">
+                            <div class="col-sm-3">
+                                <p class="mb-0">Street Address</p>
+                            </div>
+                            <div class="col-sm-9">
+                                <p class="text-muted mb-0" th:text="*{streetAddress}"></p>
+                            </div>
+                        </div>
+                        <hr>
+                        <div class="row">
+	                        <div class="col-sm-3">
+		                        <p class="mb-0">City and State</p>
+	                        </div>
+	                        <div class="col-sm-9">
+		                        <p class="text-muted mb-0" th:text="*{city + ', ' + state}"></p>
+	                        </div>
+                        </div>
+                        <hr>
+                        <div class="row">
+	                        <div class="col-sm-3">
+		                        <p class="mb-0">Zip Code</p>
+	                        </div>
+	                        <div class="col-sm-9">
+		                        <p class="text-muted mb-0" th:text="*{zipCode}"></p>
+	                        </div>
+                        </div>
+                        <hr>
+                        <div class="row">
+                            <div class="col-sm-3">
+                                <p class="mb-0">Username</p>
+                            </div>
+                            <div class="col-sm-9">
+                                <p class="text-muted mb-0" th:text="*{username}"></p>
+                            </div>
+                        </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+
+<!-- Optional JavaScript -->
+<!-- jQuery first, then Popper.js, then Bootstrap JS -->
+<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+</body>
+</html>
